feat ✨: Implemented RESTful API for managing item units with CRUD operations and refactored services for better database interactions and error handling.
- Implemented RESTful API for managing item units with create, read, update, delete, and list operations.
- Updated executable file with new build.
- Added a new executable file to the project.
- Refactored ConditionLookupService to simplify and standardize CRUD operations.
- Refactored ItemUnitService to improve database interactions and error handling.
- Updated data types of id parameter in GetByID, Delete, and Create methods to uint from uint8 for better compatibility.
- Updated data type of id parameter in GetByID and Delete functions from int8 to uint.
- Updated data type for ID from uint8 to uint in StatusLookupService methods.

+++ services/lending_type_lookup_service.go
@@ -21,7 +21,7 @@ func (s *LendingTypeLookupService) Create(lendingType *models.LendingTypeLookup)
     return err
 }
 
-func (s *LendingTypeLookupService) GetByID(id uint8) (*models.LendingTypeLookup, error) {
+func (s *LendingTypeLookupService) GetByID(id uint) (*models.LendingTypeLookup, error) {
     lendingType := &models.LendingTypeLookup{IdType: id}
     err := s.ormer.Read(lendingType)
     if err == orm.ErrNoRows {
@@ -29,7 +29,6 @@ func (s *LendingTypeLookupService) GetByID(id uint8) (*models.LendingTypeLookup,
     }
     return lendingType, err
 }
-
 func (s *LendingTypeLookupService) GetByTypeName(typeName string) (*models.LendingTypeLookup, error) {
     lendingType := &models.LendingTypeLookup{TypeName: typeName}
     err := s.ormer.Read(lendingType, "TypeName")
@@ -47,7 +46,7 @@ func (s *LendingTypeLookupService) Update(lendingType *models.LendingTypeLookup)
     return err
 }
 
-func (s *LendingTypeLookupService) Delete(id uint8) error {
+func (s *LendingTypeLookupService) Delete(id uint) error {
     lendingType := &models.LendingTypeLookup{IdType: id}
     _, err := s.ormer.Delete(lendingType)
     return err
